Country entity

+++ src/main/java/com/ribbonmix/core/repository/CountryRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * Created by Amer on 7/29/2016.
  */
-public interface CountryRepository extends JpaRepository<CountryEntity, Long> {
+public interface CountryRepository extends JpaRepository<CountryEntity, Integer> {
 
     CountryEntity findOneByIso2(String iso2);
 }
